ci2: Move cron schedule into generator.yml
Currently, the only tree we will build more often is mainline because
the pushes are sporatic. -next rebuilds on weekdays because it is
updated M-F and stable is only updated at max every few days. We want to
keep stable daily to catch LLVM changes. There could be a case for
building some of the Android trees more rapidly but we will deal with
that when we enable them.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>