generator.yml: Add list of targets to build
We primarily need this due to a lack of running 'make dtbs' on PowerPC
on 4.14 and 4.19:

https://builds.tuxbuild.com/1nUz5vGrfZUHiNGS9vtAukVBK6s/
https://builds.tuxbuild.com/1nUx2XXh3IgqIVijCac9Zn6M6zc/

This was fixed in 4.20 but it does not seem like this is suitable to
backport to 4.19 and 4.14 because the commit explicitly mentions
potentially breaking setups:

https://git.kernel.org/linus/1acf1cf8638a801e67666425d8663f813e1dae8a
https://git.kernel.org/linus/27e88af4e94ee853261972aca1a364c4c567123d

This is easy enough to work around in build-set with the targets list.
It makes things a little bit more efficient anyways because we only need
dtbs for the arm32_v5 and arm32_v6 targets; they are unused on every
other architecture. Save the cycles.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>