generate_workflow.py: Account for defconfig lists
TuxSuite has support for specifying a list of configuration options, for
example:

config: [defconfig, CONFIG_EFI=n]

which will use defconfig as a base then merge CONFIG_EFI=n into it. This
will be needed for RISC-V LLVM_IAS=1 support.

Currently, generate_workflow.py will not handle this gracefully.
Introduce get_config_name, which will print the "base config" for the
CONFIG environment variable, and print_config_name, which will print the
base config name and any options that were disabled so that it is
obvious to the viewer what configuration was used.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>