ci2: Add support for all*configs
It is advantageous to add support for all*config targets, as they will
help validate clang and LLVM further. With this change, we add a
separate build set when all*configs are detected in the matrix. Having a
separate build set will allow us to have two separate tuxsuite
instances, one for defconfigs (which should be pretty quick) and one for
all*configs, which can afford to take a while.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>