ci2: Find build by full set of configuration options
Currently, the way that we find builds is a little fragile: we just rely
on the first configuration value, not any subsequent ones. We are about
to start testing multi_v7_defconfig with CONFIG_THUMB2_KERNEL=y, which
means there will be two multi_v7_defconfig builds. To get the right
build, we need to factor in the additional configuration options. To do
that, we will just concatenate the options together with a "," between
then.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>