ci2: Completely rework YAML generation
As it stands now, we generate the TuxSuite files manually and generate
the workflow files from a fragemnt then edit it.

Instead, unify the way that YAML is generated, relying on lists and
dictionaries. This will allow us to easily generate multiple build sets
and handle conditionally adding build set values such as custom image
targets.

This should have no functional change because we are just renaming and
restructuring the YAML.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>