generator.yml: Temporarily move Android arm32 allmodconfig LLVM 13 builds to LLVM=1
The fix for https://github.com/ClangBuiltLinux/linux/issues/1355 has
been merged into mainline but due to the nature of the merge window and
the stable trees, we will not see that patch in the stable trees for at
least two weeks (after 5.13-rc1). To move the build back to green,
disable the integrated assembler for these files. As the patch trickles
down into the Android trees, we can re-enable this.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>