ci: Get max LLVM version from LLVM_TOT_VERSION
Currently, we assume that the LLVM tip of tree version is the max
version in the llvm_versions section of generator.yml and it is
hardcoded at the top of utils.py.

The addition of clang-android is going to break the first assumption and
we should avoid hardcoding things so use the newly created
LLVM_TOT_VERSION to get the max version in these places.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>