generate_workflow.py: Use unique name for output_artifact
We have two TuxSuite jobs, one for defconfigs and one for all*configs,
so that we can start testing the defconfigs while the all*configs still
build. Even though the TuxSuite jobs happen at the same time, they are
usually serial due to the large difference in the number of file being
built.

However, if a TuxSuite job were to finish before all of the boot testing
of the other TuxSuite job was finished, the output_artifact would get
overwritten and cause all of the previous jobs to fail.

To avoid this race, cache the builds.json under a slightly more
descriptive name using the build set name as a suffix.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>