check_logs.py: Do not install dependencies when BOOT=0
It is a waste of resources/time to install dependencies when we are not
going to boot the image, either because there were errors or we are not
expecting the image to boot (BOOT=0). Only install the dependencies
right before we boot. This should result in a lot fewer failed builds
due to 'apt-get update' failing.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>