generator.yml: Add support for arm64's compat vDSO
Set CROSS_COMPILE_COMPAT to arm-linux-gnueabihf- so that the compat vDSO
gets built automatically.

CONFIG_COMPAT_VDSO depends on !CPU_BIG_ENDIAN so it is not set for arm64
big endian builds. CONFIG_COMPAT_VDSO is under the CONFIG_COMPAT menu,
which is not enabled for allnoconfig so it is not set for allnoconfig
build either.

Closes: https://github.com/ClangBuiltLinux/continuous-integration2/issues/133
Signed-off-by: Nathan Chancellor <nathan@kernel.org>