utils.py: Do not be clever about distro maps
Rather than checking if the entry exists in the map and using the value
of the architecture in the URL if not, just add the entries to the map
and always return from the map. This is a little bit more verbose but
there will be a more obvious error if we add an architecture without
supporting it here.

Suggested-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Nathan Chancellor <nathan@kernel.org>