generate_{tuxsuite,workflow}.py: Add support for seperate distribution configs build set
Right now, the distribution configurations are lumped in with the
all*configs set, which means that we cannot get boot results until all
of the all*configs are done building.

Separate out the distribution configs into their own set so that they
can be built then boot tested in parallel with the defconfigs and
allconfigs build sets.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>