check_logs.py: Increase timeout to 15m for CONFIG_KASAN builds
clang-11 and clang-12 builds with CONFIG_KASAN_SW_TAGS=y are incredibly
slow to boot up but clang-13 and newer builds are generally fine. There
has been some flakiness around CONFIG_KASAN_VMALLOC=y as well.

Increase the timeout to 15m if CONFIG_KASAN=y is in the list of configs
so that we are sure that the build timed out rather than just being
slow, as GitHub Action runners are virtualized and dual core, meaning we
are paying a heavy penalty when using QEMU's TCG backend.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>