generator.yml: Disable -Werror for all{mod,yes}config
CONFIG_WERROR is enabled by default when CONFIG_COMPILE_TEST is enabled.
Disable it explicitly so we can catch actual build breakage.

I am of the opinion we should prioritize getting warning clean so that
we can re-enable this, as this does not affect our boot tests anymore
but will make it obvious when there are potential issues due to
warnings.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>