ci: Switch to directory-based patch management
TuxSuite supports supplying patches in a directory, which allows us to
break them apart and manage them easier. This requires a series file
that lists the patches to be applied.

Update the README to describe how to update the patches most
efficiently. A future commit will check the patches in CI to make sure
they are updated and the series file is valid on pull requests.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>