problem-matchers: Add kernel Oops problem matcher
Flag Oopses as warnings. For example:

[   19.442838] BUG: KASAN: slab-out-of-bounds in kmalloc_oob_right+0x1e1/0x5e0
[   19.443092] Write of size 1 at addr ffff88800134ce73 by task kunit_try_catch/87
[   19.443289]
[   19.445823] cfg80211: failed to load regulatory.db
[   19.446723] CPU: 2 PID: 87 Comm: kunit_try_catch Not tainted 5.15.0-rc2-next-20210922 #1
[   19.448836] Hardware name: QEMU Standard PC (i440FX + PIIX, 1996), BIOS 1.13.0-1ubuntu1.1 04/01/2014
[   19.450931] Call Trace:

The only reasonably common element of Oopses is the "Call Trace" part.

Additionally disable this matcher when running under
CONFIG_K?SAN_KUNIT_TEST, which intentionally triggers Oopses in order
to test KASAN, and should not be considered problems (actually LACK of
Oopses is a problem here, but we're not solving that right now).

Signed-off-by: Kees Cook <keescook@chromium.org>