check_logs.py: Register all problem matchers then remove oops matcher conditionally
We currently call the add-matcher toolkit command on an individual
basis, leaving the opportunity to forget to add a call to the
add-matcher command when a new matcher is added.

We can fix this by registering all matchers at first then disabling the
oops matcher with the remove-matcher command when using a sanitizer
build.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>