check_logs.py: Print clang date and git hash from metadata.json
It is useful when triaging a build or boot breakage to see the date that
clang was checked out and the hash that it was checked out at. Debian's
clang has this information but it is hard to parse quickly.

Use parse-debian-clang.sh to do this for clang-nightly, as that is the
only version of clang that moves fast enough to justify it.

Closes: https://github.com/ClangBuiltLinux/continuous-integration2/issues/220
Signed-off-by: Nathan Chancellor <nathan@kernel.org>