ci: Restructure check-clang.sh
Currently, check-clang.sh only checks that clang was built within the
past five days. A follow up change will add the ability to print the
info about the toolchain, such as the date it was built in a more human
readable format at the hash that it was built at.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>