generator.yml: Remove modules target
Currently, we build the config, kernel, and modules target for all
configurations and dtbs for the couple of machines that we need them
for.

The kernel and modules target for tuxsuite is not instructing it to
build the kernels and modules; instead, the kernel and modules targets
tell tuxsuite to make the kernel image and modules available for
download respectively. The default target, which is run as a
prerequisite for both of these targets, is what actually builds the
kernel image and modules. We do not care about the modules being
available after the fact, as we do not use them for anything; we only
care that they build.

Get rid of the modules target altogether, which may save us from
timeouts related to tarring up the modules for download.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>