markdown-badges.sh: Fix badge syntax
Per GitHub's own documentation, the syntax for the badge image is

https://github.com/<OWNER>/<REPOSITORY>/actions/workflows/<WORKFLOW_FILE>/badge.svg

Currently, ours is

https://github.com/<OWNER>/<REPOSITORY>/workflows/<WORKFLOW_FILE_MINUS_EXTENSION>/badge.svg

To follow GitHub's own documentation, use the filename plus extension.

After the split of the workflow files and the renaming in #305, the
links will no longer show just the workflow in question. Fix this by
using the workflow file link instead of running a query.

Link: https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/adding-a-workflow-status-badge
Signed-off-by: Nathan Chancellor <nathan@kernel.org>