generator.yml: Test certain Linux + clang combinations less often
For trees that do not move as frequently, such as stable and the Android
trees, there is little point to testing every day with versions of clang
that are EOL, since they are not being updated anymore.

To help load balance, build the stable trees on Wednesday and the
Android trees on Sunday.

Closes: https://github.com/ClangBuiltLinux/continuous-integration2/issues/200
Signed-off-by: Nathan Chancellor <nathan@kernel.org>