patches: Add an objtool patch for 5.15 and latest stable
This objtool patch is now merged into mainline and it is tagged for
stable but stable patches that are merged during the merge window tend
to have a minor delay getting into stable due to the chaos of the merge
window.

Add this patch so that the builds can go back to green (barring any LLVM
regressions...).

Signed-off-by: Nathan Chancellor <nathan@kernel.org>