check_logs.py: Increase attempts to verify build status to 9
Even with the current max retries of 7, we still occasionally see builds
that do not have a pass/fail status. Remi recommends pinging the tuxapi
endpoint but further investigation will be needed to see if this is
feasible across the GitHub Actions job boundary. In the meantime,
increase the maximum number of retries to 9, which will result in an 8m
30s final wait and a total of 17 minutes of waiting for an updated
result.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>