generate_workflow.py: Increase 'tuxsuite plan' timeout limit
By default, GitHub Actions limits each step in a workflow to 360
minutes, which would normally be enough but if an all*config build gets
preempted and has to restart, it is possible for the 'tuxsuite plan'
command to exceed 360 minutes. Increase it by two more hours up to a
total of eight hours to see if that is enough to avoid job timeouts.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>