scripts/markdown-badges.py: Add "Check clang version" badges
This workflow failing means that the version that we are using is not
staying in sync with tip of tree LLVM, which means we are potentially
losing out on test coverage. This should be more obvious so that it can
be dealt with quicker so add it to the markdown badge generator. It does
not need to be included in the actual table so it is added to the end.

Closes: https://github.com/ClangBuiltLinux/continuous-integration2/issues/478
Signed-off-by: Nathan Chancellor <nathan@kernel.org>