generate_tuxsuite.py: Rename unused data parameter
vulture flags:

  generate_tuxsuite.py:14: unused variable 'data' (100% confidence)

We do not need to use the variable but it must be present in the
parameter list to override ignore_aliases() in yaml.SafeDumper. Make it
clear to vulture that this variable is required but unused by prefixing
the variable name with '_', which is the documented way to silence this
warning.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>