scripts: Rewrite parse-debian-clang.sh in Python
The datetime module in the Python standard library is much more robust
for what this script is looking to do, on top of using re for getting
the date and commit hash from the version string in the first place.

This allows us to print more relevant information like the actual delta
between the clang checkout date and when the script runs, which is what
we really want to know, instead of just the current date.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>