scripts: Rewrite check-matrix.sh in Python
This is almost a 1:1 rewrite but it takes the opportunity to improve on
the output of the script, such that it just reports the workflows that
are over the limit and the number of jobs they have. This was way harder
to do in bash but now that we have actual data structures, it is rather
simple.

Additionally, the "nothing to report" case is much faster with the
Python rewrite.

  Benchmark 1: scripts/check-matrix.sh
    Time (mean ± σ):     871.6 ms ±   5.4 ms    [User: 569.5 ms, System: 314.6 ms]
    Range (min … max):   861.3 ms … 879.8 ms    10 runs

  Benchmark 2: scripts/check-matrix.py
    Time (mean ± σ):      33.4 ms ±   7.9 ms    [User: 28.6 ms, System: 4.6 ms]
    Range (min … max):    30.0 ms …  58.6 ms    55 runs

  Summary
    'scripts/check-matrix.py' ran
     26.09 ± 6.14 times faster than 'scripts/check-matrix.sh'

Signed-off-by: Nathan Chancellor <nathan@kernel.org>