patches: mainline: Add tentative fix for 99cb0d917ff
Commit 99cb0d917ffa ("arch: fix broken BuildID for arm64 and riscv")
exposed a GNU ld bug that is present in the version that tuxsuite uses.
We still rely on GNU ld for at least powerpc and s390, so this issue
breaks those builds.

Add Ard's tentative fix for the issue so our builds go back to green,
allowing us to easily catch other regressions.

Link: https://lore.kernel.org/CAMj1kXHqQoqoys83nEp=Q6oT68+-GpCuMjfnYK9pMy-X_+jjKw@mail.gmail.com/
Signed-off-by: Nathan Chancellor <nathan@kernel.org>