generator.yml: Drop s390 builds for LLVM < 13 on 5.10
Commit 82d3edb50a11 ("s390/cpum_sf: add READ_ONCE() semantics to compare
and swap loops") was applied to 5.10, where it started causing build
breakage with LLVM 11 and 12:

  arch/s390/kernel/perf_cpum_sf.c:1234:3: error: inline asm error: This value type register class is not natively supported!
                  "       cdsg    %[old],%[new],%[ptr]\n"
                  ^
  warning: Compiler has made implicit assumption that TypeSize is not scalable. This may or may not lead to broken code.

The construct that is used in the inline assembly was only supported in
LLVM starting with the 13.0.0 release, even though it has been supported
in GCC for a while. This was the catalyst for commit e2bc3e91d91e
("scripts/min-tool-version.sh: Raise minimum clang version to 13.0.0 for
s390").

The min-tool-version.sh infrastructure does not exist in 5.10 and it may
be undesirable to backport it. Do what was done in #150 and #161 and
disable LLVM 11 and 12 builds on stable-5.10. This does not lose us much
coverage, as we were only testing one configuration on these branches.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>