Add a script to build TuxSuite configurations locally
Sometimes, it may be desirable to perform the same set of builds that
TuxSuite does locally, such as validating that a local tree does not
have any problems.

TuxMake is the backend of TuxSuite and it is written in Python, so we
can import it, parse the TuxSuite YAML files ourselves, then offload all
the builds onto TuxMake.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>