Update 64-bit big endian PowerPC configuration target
A patch in review will change pseries_defconfig from a big endian
configuration to a little endian configuration:

https://lore.kernel.org/20230414132415.821564-32-mpe@ellerman.id.au/

As suggested in the commit message, switch to ppc64_guest_defconfig now
to guarantee that we always get a 64-bit big endian kernel that can boot
in QEMU, which is exactly what this configuration intends to do,
regardless of whether or not that change is actually accepted.

This configuration target is available on 5.4, the oldest kernel we
build ppc64 on, and it works fine with both clang-12 and clang-17, the
oldest and newest toolchain versions that this configuration is built
with.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>