ci: Use pathlib and absolute paths everywhere
Many of the paths in continuous-integration2 are relative, which can
make it difficult to verify that files are placed in the correct place.

To make this a little more ergonomic to work with, switch exclusively to
pathlib and use absolute paths to all files. This allows us to
deduplicate and simplify the code in a few places using pathlib's
methods, which are a little higher level than the os functions used in a
few places.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>