ci: Wire up '--gh-json-file'
This will prevent us from hitting GitHub API rate limits when fetching
rootfs images within boot-utils, as we will cache the release JSON as a
file during the TuxSuite step and pass it along to all the boot jobs
like we do builds.json.

The curl command is simple enough that it could be included in
generate_workflow.py directly but PyYAML makes printing it somewhat
complicated because of the quotations that are needed around the
headers. Having a separate script is less brittle for a few extra lines
of code.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>