scripts/build-local.py: Improve interrupt behavior
Currently, when interrupting scripts/build-local.py with Ctrl-C, the
current build will be killed then the next one will proceed. When lots
of builds are queued, it can take a long time to Ctrl-C them all so that
the script can finish.

Add an interrupt handler in build-local.py to immediately exit when
Ctrl-C is pressed.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>