TEMPORARY: generator.yml: Delay LLVM ToT builds until next Tuesday
There is currently a crash and infinite loop in clang's code generation,
which causes builds to fail or indefinitely hang. Due to the long
Thanksgiving weekend in the US, it is unlikely this issue will be fixed
before next week. Delay the LLVM tip of tree builds until Tuesday,
November 28th, so that TuxSuite is not tied up with known broken builds.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>