generator.yml: Update LoongArch builds for most recent developments
CONFIG_RELOCATABLE has been fixed since 6.7 with two changes (the first
one fixes the build and the second one fixes boot):

  aa0cbc1b506b ("LoongArch: Record pc instead of offset in la_abs relocation")
  eea673e9d5ea ("LoongArch: Apply dynamic relocations for LLD")

Re-enable CONFIG_RELOCATABLE (removing CONFIG_CRASH_DUMP=n as well, as
it was only disabled to allow CONFIG_RELOCATABLE to be disabled) for all
trees, as those changes have also been backported to linux-6.6.y.

The minimum supported version of clang for LoongArch has been updated to
18.0.0 in -next with commit 4d35d6e56447 ("scripts/min-tool-version.sh:
Raise minimum clang version to 18.0.0 for loongarch") because it is the
first version that supports __attribute__((model("extreme"))), which
allows us to build with CONFIG_MODULES.

In order to support CONFIG_MODULES unconditionally (i.e., not disable it
at all), we must drop support for clang-17 everywhere. We already have
to do it for -next and mainline due to the aforementioned commit. Do so,
which helps drastically simplify the matrix.

Because CONFIG_MODULES does not need to be disabled anymore, we can
build allmodconfig instead of allyesconfig, which also allows us to
stop disabling CONFIG_KCOV because the image size is much smaller.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>