caching: Add hash of contents of patches folder
Currently, the frontend caching infrastructure is unaware of the patches
being applied. This is problematic if we have changed either added,
removed, or changed any patches since the previously cached run because
the caching would skip the build even when the build's expected outcome
would be different than the cached build's result.

Hash the contents of the patches folder using the hashlib library and
add that to the cache so that when the patches folder changes at all,
the cache is skipped and the build proceeds.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>