patches: Patch error between '-msoft-float' and '-maltivec'
This is due to a change on the clang side. A patch has been submitted to
fix Linux but it will take some time to filter through all trees. To
make it easier to spot other regressions, apply the fix, which will be
dropped as it is merged into the various trees.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>