generator: yml: Use more LLVM utilities for s390 builds on linux-next with LLVM tip of tree
s390 support in ld.lld has recently landed in the main branch of LLVM.
With this, it makes sense to start cutting s390 over to full LLVM=1,
versus the current situation of just CC=clang.  linux-next is the only
tree that currently has the necessary changes to support ld.lld but they
will trickle down to other trees in time.

llvm-objcopy currently works on LLVM main as well but that needs a
kernel side fix to boot properly, so we will stick with GNU objcopy
until that is fixed.

Closes: https://github.com/ClangBuiltLinux/continuous-integration2/issues/710
Link: https://github.com/ClangBuiltLinux/linux/issues/1996
Link: https://github.com/llvm/llvm-project/commit/fe3406e349884e4ef61480dd0607f1e237102c74
Link: https://github.com/llvm/llvm-project/commit/3c02cb7492fc78fb678264cebf57ff88e478e14f
Signed-off-by: Nathan Chancellor <nathan@kernel.org>