generator: yml: Use LLVM=1 for ARCH=s390 on linux-next with tip of tree LLVM
With LLVM main, which has the necessary changes to ld.lld and
llvm-objcopy to support s390, and linux-next, which has the necessary
changes to the kernel to support these tools, ARCH=s390 builds cleanly
with LLVM=1. Cut builds that satisfy these conditions over to LLVM=1 so
that regressions with this working setup do not occur.

Link: https://github.com/llvm/llvm-project/commit/fe3406e349884e4ef61480dd0607f1e237102c74
Link: https://github.com/llvm/llvm-project/commit/3c02cb7492fc78fb678264cebf57ff88e478e14f
Link: https://github.com/ClangBuiltLinux/linux/issues/859
Link: https://github.com/ClangBuiltLinux/linux/issues/1991
Link: https://github.com/ClangBuiltLinux/linux/issues/1992
Link: https://github.com/ClangBuiltLinux/linux/issues/1993
Link: https://github.com/ClangBuiltLinux/linux/issues/1994
Link: https://github.com/ClangBuiltLinux/linux/issues/1995
Link: https://github.com/ClangBuiltLinux/linux/issues/1996
Signed-off-by: Nathan Chancellor <nathan@kernel.org>