generator: workflow: Use virtual environment for package installation
Recently, the check cache stage started failing with:

  error: externally-managed-environment

  × This environment is externally managed
  ╰─> To install Python packages system-wide, try apt install
      python3-xyz, where xyz is the package you are trying to
      install.

      If you wish to install a non-Debian-packaged Python package,
      create a virtual environment using python3 -m venv path/to/venv.
      Then use path/to/venv/bin/python and path/to/venv/bin/pip. Make
      sure you have python3-full installed.

      If you wish to install a non-Debian packaged Python application,
      it may be easiest to use pipx install xyz, which will manage a
      virtual environment for you. Make sure you have pipx installed.

      See /usr/share/doc/python3.11/README.venv for more information.

This occurs because the container this command is being run in got
upgraded silently from Debian 11 to Debian 12, where the newer Python
version allows distributors to block using pip on the system level by
default.

To avoid this altogether, use a virtual environment to ensure that
package installation works no matter what version of Debian or Ubuntu is
being used.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>