generator: yml: Add and enable Fedora LTO builds
Fedora builds with LTO when building with clang upstream, which is an
important difference from how we currently build their configuration.
Having side by side builds of LTO and no LTO is valuable for quickly
narrowing down issues caused or exposed by LTO. Add a Fedora + ThinLTO
build to match the official configuration and catch any regressions with
this configuration from LLVM.

This is enabled for stable (6.8) and newer with LLVM 17 and newer to
avoid adding too many builds to the matrix, which matches the versions
of LLVM and the kernel that are in Fedora 40 and newer. This adds 90
builds per week.

Closes: https://github.com/ClangBuiltLinux/linux/issues/2009
Signed-off-by: Nathan Chancellor <nathan@kernel.org>