generator: Handle subsystem specific -Werror configurations
Certain subsystems such as DRM have introduced configurations that
enable -Werror for their subsystem specifically, rather than rely on the
global CONFIG_WERROR because it is too big of a hammer, allowing issues
in other subsystems out of their control to break their builds.

Unfortunately, due to the way Kconfig works, CONFIG_WERROR=n does not
imply 'n' for other subsystem -Werror configurations automatically, so
the current situation of CONFIG_WERROR=n to avoid -Werror altogether
does not work.

Luckily, the generator can look at the list of configurations to see if
CONFIG_WERROR=n is present and automatically add the subsystem specific
-Werror configurations. Currently, only CONFIG_DRM_WERROR is the only
known problematic ones, so that is the only one that is disabled right
now, but others may be added later, so keep it extensible.

While CONFIG_DRM_WERROR only exists in -next currently, it does not hurt
to add it to every tree unconditionally for simplicity's sake.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>